# payment/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bayarin API",
        "description": "Hanya Untuk Patner",
        "version": "1.0.0"
    },
    "paths": {
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get List of Transactions",
                "description": "Retrieve a list of transactions for a specific wallet.",
                "operationId": "getTransactions",
                "parameters": [
                    {
                        "name": "wallet_id",
                        "in": "query",
                        "description": "ID of the wallet to get transactions for.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of transactions retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "default"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid wallet ID provided"
                    },
                    "404": {
                        "description": "Wallet not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Create a new transaction",
                "description": "Store a new transaction in the system.",
                "operationId": "createTransaction",
                "requestBody": {
                    "description": "Transaction data to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "wallet_id",
                                    "type",
                                    "amount",
                                    "date",
                                    "status"
                                ],
                                "properties": {
                                    "wallet_id": {
                                        "description": "ID of the wallet",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "type": {
                                        "description": "Type of transaction",
                                        "type": "string",
                                        "example": "deposit"
                                    },
                                    "amount": {
                                        "description": "Amount of the transaction",
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "date": {
                                        "description": "Date of the transaction",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-01"
                                    },
                                    "description": {
                                        "description": "Description of the transaction",
                                        "type": "string",
                                        "example": "Payment for service"
                                    },
                                    "recipient_wallet_id": {
                                        "description": "ID of the recipient wallet (for transfers)",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "status": {
                                        "description": "Status of the transaction",
                                        "type": "string",
                                        "example": "completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/transactions/{id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get transaction details",
                "description": "Retrieve the details of a specific transaction by ID.",
                "operationId": "getTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the transaction to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Update an existing transaction",
                "description": "Update the details of an existing transaction.",
                "operationId": "updateTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the transaction to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated transaction data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status",
                                    "date"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Status of the transaction",
                                        "type": "string",
                                        "example": "completed"
                                    },
                                    "description": {
                                        "description": "Description of the transaction",
                                        "type": "string",
                                        "example": "Updated payment details"
                                    },
                                    "date": {
                                        "description": "Updated date of the transaction",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-02"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "404": {
                        "description": "Transaction not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Delete a transaction",
                "description": "Delete an existing transaction from the system.",
                "operationId": "deleteTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the transaction to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction deleted successfully"
                    },
                    "404": {
                        "description": "Transaction not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Transaction": {
                "properties": {
                    "id": {
                        "description": "Transaction ID",
                        "type": "string",
                        "example": "12345"
                    },
                    "wallet_id": {
                        "description": "ID of the wallet associated with the transaction",
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "description": "Type of transaction",
                        "type": "string",
                        "example": "deposit"
                    },
                    "amount": {
                        "description": "Amount of money involved in the transaction",
                        "type": "integer",
                        "example": 100
                    },
                    "date": {
                        "description": "Date of the transaction",
                        "type": "string",
                        "format": "date",
                        "example": "2024-09-01"
                    },
                    "status": {
                        "description": "Current status of the transaction",
                        "type": "string",
                        "example": "completed"
                    },
                    "description": {
                        "description": "Description of the transaction",
                        "type": "string",
                        "example": "Payment for service"
                    },
                    "recipient_wallet_id": {
                        "description": "ID of the recipient wallet (for transfers)",
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BasicAuth": {
                "type": "http",
                "description": "Basic Authentication using username and password",
                "scheme": "basic"
            }
        }
    },
    "tags": [
        {
            "name": "Transactions",
            "description": "Transactions"
        }
    ]
}